requirements.txt
aisuite==0.1.6
python-dotenv==1.0.1